+++ lib/cocoapods/installer/xcode/pods_project_generator.rb
@@ -11,14 +11,11 @@ module Pod
         require 'cocoapods/installer/xcode/pods_project_generator/pod_target_installer'
         require 'cocoapods/installer/xcode/pods_project_generator/file_references_installer'
         require 'cocoapods/installer/xcode/pods_project_generator/aggregate_target_installer'
+        require 'cocoapods/installer/xcode/pods_project_generator_result'
 
         # @return [Sandbox] The sandbox where the Pods should be installed.
         #
         attr_reader :sandbox
-
-        # @return [Pod::Project] the `Pods/Pods.xcodeproj` project.
-        #
-        attr_reader :project
 
         # @return [Array<AggregateTarget>] The model representations of an
         #         aggregation of pod targets generated for a target definition
@@ -62,28 +59,44 @@ module Pod
         end
 
         def generate!
-          prepare
-          install_file_references
-          @target_installation_results = install_targets
-          integrate_targets(@target_installation_results.pod_target_installation_results)
-          wire_target_dependencies(@target_installation_results)
+          project = prepare
+          install_file_references(project)
+          target_installation_results = install_targets(project)
+          integrate_targets(target_installation_results.pod_target_installation_results)
+          wire_target_dependencies(project, target_installation_results)
           project.development_pods.children.objects.each do |group|
             group.set_path(group.real_path.realpath)
           end
-          @target_installation_results
+          PodsProjectGeneratorResult.new(project, target_installation_results)
         end
 
-        def write
-          UI.message "- Writing Xcode project file to #{UI.path sandbox.project_path}" do
+        # Writes the project to the provided path, performing any final steps necessary before saving
+        # the project to disk
+        #
+        # @param [Pod::Project] project the project to write
+        #
+        # @param [InstallationResults] target_installation_results
+        #        the installation results to use when creating schemes
+        #
+        # @param [Pathname] destination_path
+        #        the path to which the project will be written
+        #
+        # @param [Boolean] deterministic_uuids
+        #        Whether to use deterministic UUIDs in the project. See {Xcodeproj#predictabilize_uuids}
+        #
+        # @return [void]
+        #
+        def self.write(project, target_installation_results, destination_path, deterministic_uuids)
+          UI.message "- Writing Xcode project file to #{UI.path destination_path}" do
             project.pods.remove_from_project if project.pods.empty?
             project.development_pods.remove_from_project if project.development_pods.empty?
             project.sort(:groups_position => :below)
-            if installation_options.deterministic_uuids?
+            if deterministic_uuids
               UI.message('- Generating deterministic UUIDs') { project.predictabilize_uuids }
             end
             library_product_types = [:framework, :dynamic_library, :static_library]
 
-            pod_target_installation_results = @target_installation_results.pod_target_installation_results
+            pod_target_installation_results = target_installation_results.pod_target_installation_results
             results_by_native_target = Hash[pod_target_installation_results.map do |_, result|
               [result.native_target, result]
             end]
@@ -96,7 +109,7 @@ module Pod
                 scheme.add_test_target(test_native_target)
               end
             end
-            project.save
+            project.save(destination_path)
           end
         end
 
@@ -104,21 +117,27 @@ module Pod
         #
         # @return [void]
         #
-        def share_development_pod_schemes
-          development_pod_targets.select(&:should_build?).each do |pod_target|
-            next unless share_scheme_for_development_pod?(pod_target.pod_name)
+        def share_development_pod_schemes(project)
+          targets = development_pod_targets.select do |target|
+            target.should_build? && share_scheme_for_development_pod?(target.pod_name)
+          end
+          targets.each do |pod_target|
             Xcodeproj::XCScheme.share_scheme(project.path, pod_target.label)
-            if pod_target.contains_test_specifications?
-              pod_target.test_specs.each do |test_spec|
-                Xcodeproj::XCScheme.share_scheme(project.path, pod_target.test_target_label(test_spec))
-              end
+            pod_target.test_specs.each do |test_spec|
+              Xcodeproj::XCScheme.share_scheme(project.path, pod_target.test_target_label(test_spec))
+            end
+
+            pod_target.app_specs.each do |app_spec|
+              Xcodeproj::XCScheme.share_scheme(project.path, pod_target.app_target_label(app_spec))
             end
           end
         end
 
-        private
-
+        # @!attribute [Hash{String => TargetInstallationResult}] pod_target_installation_results
+        # @!attribute [Hash{String => TargetInstallationResult}] aggregate_target_installation_results
         InstallationResults = Struct.new(:pod_target_installation_results, :aggregate_target_installation_results)
+
+        private
 
         def create_project
           if object_version = aggregate_targets.map(&:user_project).compact.map { |p| p.object_version.to_i }.min
@@ -130,37 +149,37 @@ module Pod
 
         # Creates the Pods project from scratch if it doesn't exists.
         #
-        # @return [void]
+        # @return [Project]
         #
         # @todo   Clean and modify the project if it exists.
         #
         def prepare
           UI.message '- Creating Pods project' do
-            @project = create_project
+            project = create_project
             analysis_result.all_user_build_configurations.each do |name, type|
-              @project.add_build_configuration(name, type)
+              project.add_build_configuration(name, type)
             end
             # Reset symroot just in case the user has added a new build configuration other than 'Debug' or 'Release'.
-            @project.symroot = Pod::Project::LEGACY_BUILD_ROOT
+            project.symroot = Pod::Project::LEGACY_BUILD_ROOT
 
             pod_names = pod_targets.map(&:pod_name).uniq
             pod_names.each do |pod_name|
               local = sandbox.local?(pod_name)
               path = sandbox.pod_realdir(pod_name)
-              @project.add_pod_group(pod_name, path, local, false)
+              project.add_pod_group(pod_name, path, local, false)
             end
 
             if config.podfile_path
-              @project.add_podfile(config.podfile_path)
+              project.add_podfile(config.podfile_path)
             end
 
-            sandbox.project = @project
+            sandbox.project = project
             platforms = aggregate_targets.map(&:platform)
             osx_deployment_target = platforms.select { |p| p.name == :osx }.map(&:deployment_target).min
             ios_deployment_target = platforms.select { |p| p.name == :ios }.map(&:deployment_target).min
             watchos_deployment_target = platforms.select { |p| p.name == :watchos }.map(&:deployment_target).min
             tvos_deployment_target = platforms.select { |p| p.name == :tvos }.map(&:deployment_target).min
-            @project.build_configurations.each do |build_configuration|
+            project.build_configurations.each do |build_configuration|
               build_configuration.build_settings['MACOSX_DEPLOYMENT_TARGET'] = osx_deployment_target.to_s if osx_deployment_target
               build_configuration.build_settings['IPHONEOS_DEPLOYMENT_TARGET'] = ios_deployment_target.to_s if ios_deployment_target
               build_configuration.build_settings['WATCHOS_DEPLOYMENT_TARGET'] = watchos_deployment_target.to_s if watchos_deployment_target
@@ -168,15 +187,16 @@ module Pod
               build_configuration.build_settings['STRIP_INSTALLED_PRODUCT'] = 'NO'
               build_configuration.build_settings['CLANG_ENABLE_OBJC_ARC'] = 'YES'
             end
+            project
           end
         end
 
-        def install_file_references
+        def install_file_references(project)
           installer = FileReferencesInstaller.new(sandbox, pod_targets, project, installation_options.preserve_pod_file_structure)
           installer.install!
         end
 
-        def install_targets
+        def install_targets(project)
           UI.message '- Installing targets' do
             umbrella_headers_by_dir = pod_targets.map do |pod_target|
               next unless pod_target.should_build? && pod_target.defines_module?
@@ -185,7 +205,7 @@ module Pod
 
             pod_target_installation_results = Hash[pod_targets.sort_by(&:name).map do |pod_target|
               umbrella_headers_in_header_dir = umbrella_headers_by_dir[pod_target.module_map_path.dirname]
-              target_installer = PodTargetInstaller.new(sandbox, @project, pod_target, umbrella_headers_in_header_dir)
+              target_installer = PodTargetInstaller.new(sandbox, project, pod_target, umbrella_headers_in_header_dir)
               [pod_target.name, target_installer.install!]
             end]
 
@@ -196,7 +216,7 @@ module Pod
             end
 
             aggregate_target_installation_results = Hash[aggregate_targets.sort_by(&:name).map do |target|
-              target_installer = AggregateTargetInstaller.new(sandbox, @project, target)
+              target_installer = AggregateTargetInstaller.new(sandbox, project, target)
               [target.name, target_installer.install!]
             end]
 
@@ -204,16 +224,24 @@ module Pod
           end
         end
 
+        # @param [Hash{String => InstallationResult}] pod_target_installation_results
+        #        the installations to integrate
+        #
+        # @return [void]
+        #
         def integrate_targets(pod_target_installation_results)
           pod_installations_to_integrate = pod_target_installation_results.values.select do |pod_target_installation_result|
             pod_target = pod_target_installation_result.target
-            !pod_target_installation_result.test_native_targets.empty? || pod_target.contains_script_phases?
+            !pod_target_installation_result.test_native_targets.empty? ||
+                !pod_target_installation_result.app_native_targets.empty? ||
+                pod_target.contains_script_phases?
           end
-          unless pod_installations_to_integrate.empty?
-            UI.message '- Integrating targets' do
-              pod_installations_to_integrate.each do |pod_target_installation_result|
-                PodTargetIntegrator.new(pod_target_installation_result, installation_options).integrate!
-              end
+          return if pod_installations_to_integrate.empty?
+
+          UI.message '- Integrating targets' do
+            use_input_output_paths = !installation_options.disable_input_output_paths
+            pod_installations_to_integrate.each do |pod_target_installation_result|
+              PodTargetIntegrator.new(pod_target_installation_result, :use_input_output_paths => use_input_output_paths).integrate!
             end
           end
         end
@@ -242,7 +270,7 @@ module Pod
         #
         # @return [void]
         #
-        def wire_target_dependencies(target_installation_results)
+        def wire_target_dependencies(project, target_installation_results)
           frameworks_group = project.frameworks_group
           pod_target_installation_results_hash = target_installation_results.pod_target_installation_results
           aggregate_target_installation_results_hash = target_installation_results.aggregate_target_installation_results
@@ -298,6 +326,24 @@ module Pod
                   end
                   test_native_target.add_dependency(dependency_installation_result.native_target)
                   add_framework_file_reference_to_native_target(test_native_target, pod_target, test_dependent_target, frameworks_group)
+                end
+              end
+            end
+
+            # Wire up app native targets.
+            unless pod_target_installation_result.app_native_targets.empty?
+              pod_target_installation_result.app_specs_by_native_target.each do |app_native_target, app_specs|
+                app_dependent_targets = app_specs.flat_map { |s| pod_target.app_dependent_targets_by_spec_name[s.name] }.compact.unshift(pod_target).uniq
+                app_dependent_targets.each do |app_dependent_target|
+                  dependency_installation_result = pod_target_installation_results_hash[app_dependent_target.name]
+                  resource_bundle_native_targets = dependency_installation_result.app_resource_bundle_targets[app_specs.first.name]
+                  unless resource_bundle_native_targets.nil?
+                    resource_bundle_native_targets.each do |app_resource_bundle_target|
+                      app_native_target.add_dependency(app_resource_bundle_target)
+                    end
+                  end
+                  app_native_target.add_dependency(dependency_installation_result.native_target)
+                  add_framework_file_reference_to_native_target(app_native_target, pod_target, app_dependent_target, frameworks_group)
                 end
               end
             end